tf-serve/requirements.txt
pandas==0.22.0
tensorflow==1.7.0
tensorflow-serving-api-python3==1.7.0
